+++ projects/angular-libs-demo/src/app/app.module.ts
@@ -9,7 +9,7 @@ import { ServiceWayComponent } from './service-way/service-way.component';
 import { ServiceCodeWayComponent } from './service-code-way/service-code-way.component';
 import { OnPushComponent } from './on-push/on-push.component';
 import { EmptyComponent } from './empty.component';
-import { UploadxModule } from 'projects/ngx-file-upload/src/lib/uploadx.module';
+import { NgxFileUploadModule } from 'projects/ngx-file-upload/src/lib/ngx-file-upload.module';
 
 @NgModule({
   declarations: [
@@ -20,7 +20,7 @@ import { UploadxModule } from 'projects/ngx-file-upload/src/lib/uploadx.module';
     ServiceCodeWayComponent,
     OnPushComponent
   ],
-  imports: [RouterModule.forRoot(AppRoutes), BrowserModule, UploadxModule],
+  imports: [RouterModule.forRoot(AppRoutes), BrowserModule, NgxFileUploadModule],
   providers: [],
   bootstrap: [AppComponent]
 })
