+++ projects/angular-libs-demo/src/app/service-way/service-way.component.html
@@ -2,7 +2,7 @@
   <div class="btns">
     <div class="btn-col">
       <label class="upload-btn">+
-        <input type="file" uploadx  multiple accept="image/*,video/*" style="display:none">
+        <input type="file" ngxfileupload  multiple accept="image/*,video/*" style="display:none">
       </label>
       <button class="btn red raised" (click)="uploadAll()">Upload All</button>
       <button class="btn red raised" (click)="cancelAll()">Cancel All</button>
