--- projects/ngx-file-upload/src/lib/ngx-file-upload.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'lib-ngx-file-upload',
-  template: `
-    <p>
-      ngx-file-upload works!
-    </p>
-  `,
-  styles: []
-})
-export class NgxFileUploadComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
