--- projects/ngx-file-upload/src/lib/ngx-file-upload.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { NgxFileUploadService } from './ngx-file-upload.service';
-
-describe('NgxFileUploadService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: NgxFileUploadService = TestBed.get(NgxFileUploadService);
-    expect(service).toBeTruthy();
-  });
-});
