--- projects/ngx-file-upload/src/lib/public_api.ts
@@ -1,10 +1,0 @@
-export * from './uploadx.module';
-
-export { UploadxOptions } from './models/upload-options';
-export { UploadxControlEvent } from './models/control-event';
-export { UploadState } from './models/upload-state';
-export { UploadItem } from './models/upload-item';
-
-export { UploadxService } from './uploadx.service';
-
-export { Uploader } from './utils/uploader';
