--- projects/ngx-file-upload/src/lib/uploadx.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { UploadxDirective } from './uploadx.directive';
-
-@NgModule({
-  imports: [CommonModule],
-  declarations: [UploadxDirective],
-  exports: [UploadxDirective]
-})
-export class UploadxModule {}
