--- projects/ngx-file-upload/src/public-api.ts
@@ -1,7 +1,0 @@
-/*
- * Public API Surface of ngx-file-upload
- */
-
-export * from './lib/ngx-file-upload.service';
-export * from './lib/ngx-file-upload.component';
-export * from './lib/ngx-file-upload.module';
