+++ projects/ngx-file-upload/src/public_api.ts
@@ -1,0 +1,15 @@
+/*
+ * Public API Surface of ngx-file-upload
+ */
+
+export { UploadxOptions } from './lib/models/upload-options';
+export { UploadxControlEvent } from './lib/models/control-event';
+export { UploadState } from './lib/models/upload-state';
+export { UploadItem } from './lib/models/upload-item';
+
+export { Uploader } from './lib/utils/uploader';
+
+export { NgxFileUploadService } from './lib/ngx-file-upload.service';
+
+export { NgxFileUploadModule } from './lib/ngx-file-upload.module';
+
