+++ projects/ngx-test-utils/ng-package.json
@@ -2,6 +2,6 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/ngx-test-utils",
   "lib": {
-    "entryFile": "src/public-api.ts"
+    "entryFile": "src/public_api.ts"
   }
 }
